requirements.txt
colorama==0.4.6
pyfiglet==1.0.2
python-dotenv==1.0.1
Requests==2.32.3
